Fix build break with rustc 1.77.0

+++ services/apps/src/apps_storage.rs
@@ -207,7 +207,7 @@ impl AppsStorage {
                 stat.blocks_available()
             );
             #[allow(clippy::useless_conversion)]
-            return (stat.block_size() * stat.blocks_available()).into();
+            return (stat.block_size() * stat.blocks_available() as u64).into();
         }
         0
     }
